m_asset_poc/requirements.txt
# Core dependencies
fastapi==0.115.13
uvicorn[standard]==0.34.3
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1

# Database
asyncpg==0.30.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41

# LangChain & LangGraph
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.25
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langgraph==0.4.10
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.3
langgraph-sdk==0.1.70
typing-extensions>=4.8.0

# Vector Database
weaviate-client==4.15.4

# Monitoring & Tracing
langfuse==3.0.5

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
sqlparse>=0.4.4

# Caching
redis==6.2.0
hiredis==3.2.1

# Excel Processing
pandas==2.3.0
openpyxl==3.1.5
aiofiles==24.1.0

# Development
pytest==8.4.1
pytest-asyncio==1.0.0
jupyterlab==4.2.5
ipykernel==6.29.5

# A2A Protocol
a2a-sdk==0.2.9

# Additional dependencies
numpy==2.2.6
pytz==2025.2
tzdata==2025.2